# StageV/src/app/services/models/mission.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Project } from '../models/project';
export interface Mission {
  description?: string;
  endDate?: string;
  id?: number;
  name?: string;
  project?: Project;
  startDate?: string;
}