# StageV/src/app/services/models/project.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Mission } from '../models/mission';
export interface Project {
  clientname?: string;
  code?: string;
  id?: number;
  missions?: Array<Mission>;
  name?: string;
}