# StageV/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8181",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/api/visa-requests/update/{id}": {
      "put": {
        "tags": [
          "visa-request-controller"
        ],
        "operationId": "updateVisaRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VisaRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/update/{id}": {
      "put": {
        "tags": [
          "project-controller"
        ],
        "operationId": "updateProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/api/passports/update/{id}": {
      "put": {
        "tags": [
          "passport-controller"
        ],
        "operationId": "update",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Passport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/update/{id}": {
      "put": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "updateMission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                }
              }
            }
          }
        }
      }
    },
    "/api/visa-requests/create": {
      "post": {
        "tags": [
          "visa-request-controller"
        ],
        "operationId": "createVisaRequest",
        "parameters": [
          {
            "name": "passportId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VisaRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VisaRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/create": {
      "post": {
        "tags": [
          "project-controller"
        ],
        "operationId": "createProject",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Project"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/api/passports": {
      "get": {
        "tags": [
          "passport-controller"
        ],
        "operationId": "getAll",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Passport"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "passport-controller"
        ],
        "operationId": "create",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Passport"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/create": {
      "post": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "createMission",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Mission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                }
              }
            }
          }
        }
      }
    },
    "/api/visa-requests/{id}": {
      "get": {
        "tags": [
          "visa-request-controller"
        ],
        "operationId": "getVisaRequestById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/VisaRequest"
                }
              }
            }
          }
        }
      }
    },
    "/api/visa-requests/all": {
      "get": {
        "tags": [
          "visa-request-controller"
        ],
        "operationId": "getAllVisaRequests",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VisaRequest"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{id}": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Project"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/stats/missions": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "countMissionsByProject",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "integer",
                    "format": "int64"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/all": {
      "get": {
        "tags": [
          "project-controller"
        ],
        "operationId": "getAllProjects",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Project"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/passports/{id}": {
      "get": {
        "tags": [
          "passport-controller"
        ],
        "operationId": "getById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Passport"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "passport-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/missions/{id}": {
      "get": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "getMission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Mission"
                }
              }
            }
          }
        }
      }
    },
    "/api/missions/all": {
      "get": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "getAllMissions",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Mission"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/visa-requests/delete/{id}": {
      "delete": {
        "tags": [
          "visa-request-controller"
        ],
        "operationId": "deleteVisaRequest",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/projects/delete/{id}": {
      "delete": {
        "tags": [
          "project-controller"
        ],
        "operationId": "deleteProject",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/missions/delete/{id}": {
      "delete": {
        "tags": [
          "mission-controller"
        ],
        "operationId": "deleteMission",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/auth/delete/{username}": {
      "delete": {
        "tags": [
          "auth-controller"
        ],
        "operationId": "delete_1",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Passport": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date"
          },
          "placeOfBirth": {
            "type": "string"
          },
          "sex": {
            "type": "string"
          },
          "occupation": {
            "type": "string"
          },
          "issuingAuthority": {
            "type": "string"
          },
          "dateOfIssue": {
            "type": "string",
            "format": "date"
          },
          "dateOfExpiry": {
            "type": "string",
            "format": "date"
          },
          "passportNumber": {
            "type": "integer",
            "format": "int64"
          },
          "primaryPassport": {
            "type": "boolean"
          }
        }
      },
      "VisaRequest": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "missionPurpose": {
            "type": "string"
          },
          "dateOfMission": {
            "type": "string",
            "format": "date"
          },
          "countryOfFirstMission": {
            "type": "string"
          },
          "countryIssuingVisa": {
            "type": "string"
          },
          "travelerType": {
            "type": "string"
          },
          "passport": {
            "$ref": "#/components/schemas/Passport"
          },
          "passportNumber": {
            "type": "string"
          },
          "issueDate": {
            "type": "string",
            "format": "date"
          },
          "expiryDate": {
            "type": "string",
            "format": "date"
          },
          "issuingCountry": {
            "type": "string"
          },
          "issuedBy": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Mission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "startDate": {
            "type": "string",
            "format": "date"
          },
          "endDate": {
            "type": "string",
            "format": "date"
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          }
        }
      },
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "clientname": {
            "type": "string"
          },
          "missions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Mission"
            }
          }
        }
      }
    }
  }
}